Relocate Test Annotations
Prerequisite for Java modularizing.
