+++ test-common/src/test/java/io/github/chrimle/o2jrm/tests/GeneratedEnumTests.java
@@ -23,14 +23,14 @@ import com.google.gson.annotations.JsonAdapter;
 import com.google.gson.stream.JsonReader;
 import com.google.gson.stream.JsonWriter;
 import io.github.chrimle.o2jrm.GeneratedSource;
-import io.github.chrimle.o2jrm.annotations.TestAnnotationOne;
-import io.github.chrimle.o2jrm.annotations.TestAnnotationThree;
-import io.github.chrimle.o2jrm.annotations.TestAnnotationTwo;
 import io.github.chrimle.o2jrm.configs.ConfigOption;
 import io.github.chrimle.o2jrm.configs.Library;
 import io.github.chrimle.o2jrm.interfaces.TestInterfaceOne;
 import io.github.chrimle.o2jrm.interfaces.TestInterfaceTwo;
 import io.github.chrimle.o2jrm.models.GeneratedField;
+import io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne;
+import io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree;
+import io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo;
 import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.GeneratorConfigurationTests.ConfigOptionsTests;
 import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.GeneratorConfigurationTests.ConfigOptionsTests.AdditionalEnumTypeAnnotationsTests;
 import io.github.chrimle.o2jrm.tests.GeneratedEnumTests.GeneratorConfigurationTests.ConfigOptionsTests.UseEnumCaseInsensitiveTests;
