+++ test-java-useJakartaEe/pom.xml
@@ -244,9 +244,9 @@
                             <modelPackage>${modelPackage.default}.${execution.additionalModelTypeAnnotations}</modelPackage>
                             <configOptions>
                                 <additionalModelTypeAnnotations>
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationOne
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationTwo
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationThree
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree
                                 </additionalModelTypeAnnotations>
                             </configOptions>
                         </configuration>
@@ -261,9 +261,9 @@
                             <library>webclient</library>
                             <configOptions>
                                 <additionalModelTypeAnnotations>
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationOne
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationTwo
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationThree
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree
                                 </additionalModelTypeAnnotations>
                             </configOptions>
                         </configuration>
@@ -276,7 +276,7 @@
                         <configuration>
                             <modelPackage>${modelPackage.default}.${execution.additionalEnumTypeAnnotations}</modelPackage>
                             <configOptions>
-                                <additionalEnumTypeAnnotations>@io.github.chrimle.o2jrm.annotations.TestAnnotationOne;@io.github.chrimle.o2jrm.annotations.TestAnnotationTwo;@io.github.chrimle.o2jrm.annotations.TestAnnotationThree;</additionalEnumTypeAnnotations>
+                                <additionalEnumTypeAnnotations>@io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne;@io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo;@io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree;</additionalEnumTypeAnnotations>
                             </configOptions>
                         </configuration>
                     </execution>
@@ -289,7 +289,7 @@
                             <modelPackage>${modelPackage.default}.${execution.webclient.additionalEnumTypeAnnotations}</modelPackage>
                             <library>webclient</library>
                             <configOptions>
-                                <additionalEnumTypeAnnotations>@io.github.chrimle.o2jrm.annotations.TestAnnotationOne;@io.github.chrimle.o2jrm.annotations.TestAnnotationTwo;@io.github.chrimle.o2jrm.annotations.TestAnnotationThree;</additionalEnumTypeAnnotations>
+                                <additionalEnumTypeAnnotations>@io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne;@io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo;@io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree;</additionalEnumTypeAnnotations>
                             </configOptions>
                         </configuration>
                     </execution>
