+++ test-java/pom.xml
@@ -252,9 +252,9 @@
                             <modelPackage>${modelPackage.default}.${execution.additionalModelTypeAnnotations}</modelPackage>
                             <configOptions>
                                 <additionalModelTypeAnnotations>
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationOne
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationTwo
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationThree
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree
                                 </additionalModelTypeAnnotations>
                             </configOptions>
                         </configuration>
@@ -269,9 +269,9 @@
                             <library>webclient</library>
                             <configOptions>
                                 <additionalModelTypeAnnotations>
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationOne
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationTwo
-                                    @io.github.chrimle.o2jrm.annotations.TestAnnotationThree
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo
+                                    @io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree
                                 </additionalModelTypeAnnotations>
                             </configOptions>
                         </configuration>
@@ -284,7 +284,7 @@
                         <configuration>
                             <modelPackage>${modelPackage.default}.${execution.additionalEnumTypeAnnotations}</modelPackage>
                             <configOptions>
-                                <additionalEnumTypeAnnotations>@io.github.chrimle.o2jrm.annotations.TestAnnotationOne;@io.github.chrimle.o2jrm.annotations.TestAnnotationTwo;@io.github.chrimle.o2jrm.annotations.TestAnnotationThree;</additionalEnumTypeAnnotations>
+                                <additionalEnumTypeAnnotations>@io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne;@io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo;@io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree;</additionalEnumTypeAnnotations>
                             </configOptions>
                         </configuration>
                     </execution>
@@ -297,7 +297,7 @@
                             <modelPackage>${modelPackage.default}.${execution.webclient.additionalEnumTypeAnnotations}</modelPackage>
                             <library>webclient</library>
                             <configOptions>
-                                <additionalEnumTypeAnnotations>@io.github.chrimle.o2jrm.annotations.TestAnnotationOne;@io.github.chrimle.o2jrm.annotations.TestAnnotationTwo;@io.github.chrimle.o2jrm.annotations.TestAnnotationThree;</additionalEnumTypeAnnotations>
+                                <additionalEnumTypeAnnotations>@io.github.chrimle.o2jrm.test.annotations.TestAnnotationOne;@io.github.chrimle.o2jrm.test.annotations.TestAnnotationTwo;@io.github.chrimle.o2jrm.test.annotations.TestAnnotationThree;</additionalEnumTypeAnnotations>
                             </configOptions>
                         </configuration>
                     </execution>
