+++ test-utils/src/main/java/io/github/chrimle/o2jrm/test/annotations/TestFieldExtraAnnotationTwo.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.github.chrimle.o2jrm.annotations;
+package io.github.chrimle.o2jrm.test.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -21,5 +21,5 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface TestAnnotationThree {}
+@Target(ElementType.FIELD)
+public @interface TestFieldExtraAnnotationTwo {}
